refactor(agents): the review's findings, applied
Five reviewers went over the change. What they found, fixed:

The self-edit tools ignored the run's taint state, so text read from outside
Activepieces could steer an agent into rewriting itself with no gate, while
merely sending an email in the same turn would have asked. A tainted turn now
refuses, and says where the change can be made instead. The model can no longer
publish what it wrote, a rewrite leaves an audit row, and the note claiming a
draft is inert was false for chat: an agent chat runs the draft, so the change
lands on the next message.

The target agent is resolved once where the conversation is loaded, rather than
re-derived deep in a switch behind a policy boolean, which also drops a second
read of the same row. The two tool-name lists live in the shared package now, so
the worker offering a tool and the server allowing it cannot drift apart.

The panel moved out of the route file, which is 1092 lines lighter, into a
route-local folder, because features/ may not import from app/ and the panel
needs the builder's tool picker. Radix owns the tabs and their aria wiring, the
scroll fade owns its own state instead of re-rendering every field, the
transition says when the panel is gone instead of a hand-synced timer, one
tri-state says which panel is open instead of two booleans kept exclusive by
hand, and the sidebar hook sits in a leaf so hovering the rail no longer
re-renders the transcript.

Deleted: the test gate and two state helpers the removed screens took with
them, their tests, sixteen orphaned strings, an unreachable write-sequence
guard, and a requirements array that only ever answered one question.
