+++ packages/core/shared/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@activepieces/shared",
-  "version": "0.163.0",
+  "version": "0.164.0",
   "type": "commonjs",
   "sideEffects": false,
   "main": "./dist/src/index.js",
