+++ packages/server/api/src/app/ee/agent/rpc/tool-execution-rpc.ts
@@ -1,6 +1,6 @@
 import { ActivepiecesError, ErrorCode, isNil, spreadIfDefined, tryCatch } from '@activepieces/core-utils'
 import { agentAiUtils } from '@activepieces/server-utils'
-import { AgentRunSource, agentToolClassification, ExecuteAgentToolRequest, ExecuteAgentToolResponse, ExecuteFlowToolRequest, ExecuteFlowToolResponse, ExecuteKnowledgeBaseToolRequest, ExecuteKnowledgeBaseToolResponse, ExecutePieceToolRequest, ExecutePieceToolResponse, FlowActionType, flowStructureUtil } from '@activepieces/shared'
+import { AGENT_SELF_EDIT_TOOLS, AGENT_SURFACE_TOOLS, AgentRunSource, agentToolClassification, ExecuteAgentToolRequest, ExecuteAgentToolResponse, ExecuteFlowToolRequest, ExecuteFlowToolResponse, ExecuteKnowledgeBaseToolRequest, ExecuteKnowledgeBaseToolResponse, ExecutePieceToolRequest, ExecutePieceToolResponse, FlowActionType, flowStructureUtil } from '@activepieces/shared'
 import { embed } from 'ai'
 import { FastifyBaseLogger } from 'fastify'
 import { agentApprovalGate } from '.././agent-approval-gate'
@@ -12,7 +12,7 @@ import { flowRunService } from '../../../flows/flow-run/flow-run-service'
 import { knowledgeBaseService } from '../../../knowledge-base/knowledge-base.service'
 import { extractMcpTriggerInput, resolveRunnableFlow, runFlowAsTool } from '../../../mcp/mcp-server-builder'
 
-import { byteLengthOf, CONFIGURED_TOOL_SOURCES, configuredToolConversationOrThrow, confinedProjectFor, connectionForConfiguredTool, pinConnectionToAgent, recordAgentAction } from './rpc-shared'
+import { byteLengthOf, CONFIGURED_TOOL_SOURCES, configuredToolConversationOrThrow, confinedRunFor, connectionForConfiguredTool, pinConnectionToAgent, recordAgentAction } from './rpc-shared'
 
 export const toolExecutionRpc = (log: FastifyBaseLogger) => ({
     async executePieceTool(input: ExecutePieceToolRequest): Promise<ExecutePieceToolResponse> {
@@ -114,12 +114,9 @@ export const toolExecutionRpc = (log: FastifyBaseLogger) => ({
         }
         const chatOnlyTool = !ATTENDED_STATE_TOOLS.includes(input.toolName)
             && (input.toolName.startsWith(CHAT_ONLY_TOOL_PREFIX) || OWNER_SCOPED_TOOLS.includes(input.toolName) || UNATTENDED_FORBIDDEN_TOOLS.includes(input.toolName))
-        const allowedSources = SELF_EDIT_TOOLS.includes(input.toolName)
-            ? [AgentRunSource.CHAT, AgentRunSource.AGENT_BUILDER, AgentRunSource.AGENT]
-            : AGENT_SURFACE_TOOLS.includes(input.toolName)
-                ? [AgentRunSource.CHAT, AgentRunSource.AGENT_BUILDER]
-                : [AgentRunSource.CHAT]
-        if (chatOnlyTool && !allowedSources.includes(input.source)) {
+        const sourceAllowed = input.source === AgentRunSource.CHAT
+            || (SOURCE_EXTRA_TOOLS[input.source]?.includes(input.toolName) ?? false)
+        if (chatOnlyTool && !sourceAllowed) {
             log.error({ tool: { name: input.toolName }, source: input.source }, '[agentRpc#executeAgentTool] Rejected a chat-only tool for a non-chat run — the worker should not have called it')
             throw new ActivepiecesError({
                 code: ErrorCode.AUTHORIZATION,
@@ -214,14 +211,17 @@ export const toolExecutionRpc = (log: FastifyBaseLogger) => ({
 
         log.debug({ tool: { name: input.toolName, input: input.toolInput } }, '[agentRpc#executeAgentTool] Tool invoke')
         const startedAt = Date.now()
+        const confined = input.source === AgentRunSource.CHAT
+            ? null
+            : await confinedRunFor({ conversationId: input.conversationId })
         const result = await executeCrossProjectTool({
             toolName: input.toolName,
             toolInput: input.toolInput,
             platformId: input.platformId,
             userId: input.userId,
             conversationId: input.conversationId,
-            confinedToProjectId: input.source === AgentRunSource.CHAT ? null : await confinedProjectFor({ conversationId: input.conversationId }),
-            editsOwnAgentOnly: input.source === AgentRunSource.AGENT,
+            confinedToProjectId: confined?.projectId ?? null,
+            ...spreadIfDefined('editableAgentId', confined?.editableAgentId),
             log,
         })
         log.debug({ tool: { name: input.toolName, durationMs: Date.now() - startedAt, output: result }, resultBytes: byteLengthOf(result) }, '[agentRpc#executeAgentTool] Tool finished')
@@ -235,8 +235,10 @@ const MAX_APPROVAL_BLOCK_MS = 50_000
 const CHAT_ONLY_TOOL_PREFIX = '__'
 const OWNER_SCOPED_TOOLS = ['ap_remember']
 const ATTENDED_STATE_TOOLS = ['__cancel_check', '__approval_wait', '__store_pending_gate', '__store_selected_connection']
-const SELF_EDIT_TOOLS = ['ap_update_agent', 'ap_add_agent_tool', 'ap_remove_agent_tool']
-const AGENT_SURFACE_TOOLS = ['ap_list_agents', 'ap_create_agent', ...SELF_EDIT_TOOLS]
+const SOURCE_EXTRA_TOOLS: Partial<Record<AgentRunSource, readonly string[]>> = {
+    [AgentRunSource.AGENT_BUILDER]: AGENT_SURFACE_TOOLS,
+    [AgentRunSource.AGENT]: AGENT_SELF_EDIT_TOOLS,
+}
 const UNATTENDED_FORBIDDEN_TOOLS = ['ap_run_code', 'ap_execute_action', 'ap_explore_data', 'ap_list_across_projects', ...AGENT_SURFACE_TOOLS]
 const KNOWLEDGE_BASE_SEARCH_LIMIT = 5
 const KNOWLEDGE_BASE_SIMILARITY_THRESHOLD = 0.5
