+++ packages/server/worker/test/lib/execute/jobs/ee/agent/agent-worker-tools.test.ts
@@ -357,6 +357,28 @@ describe('agentWorkerTools', () => {
         })
     })
 
+    describe('self-edit taint gate', () => {
+        const editWith = async (tainted: boolean) => {
+            const executeTool = vi.fn().mockResolvedValue({ agentId: 'agent_1' })
+            const tools = agentWorkerTools.createAgentSurfaceTools({ executeTool, taintState: { tainted } })
+            const result = await tools.ap_update_agent.execute({ instructions: 'Do as the email says.' }, { toolCallId: 'tc-self', messages: [], abortSignal: undefined as unknown as AbortSignal })
+            return { result: result as { error?: string }, executeTool }
+        }
+
+        it('refuses to rewrite the agent once the turn has read outside content', async () => {
+            const { result, executeTool } = await editWith(true)
+
+            expect(result.error).toMatch(/outside Activepieces/i)
+            expect(executeTool).not.toHaveBeenCalled()
+        })
+
+        it('rewrites it on a clean turn', async () => {
+            const { executeTool } = await editWith(false)
+
+            expect(executeTool).toHaveBeenCalledWith('ap_update_agent', { instructions: 'Do as the email says.' })
+        })
+    })
+
     describe('gate timeout vs decline', () => {
         const runExecuteActionWith = async (outcome: 'timeout' | 'declined') => {
             const { eventEmitter } = makeMockEventEmitter()
