+++ packages/web/public/locales/en/translation.json
@@ -22,7 +22,6 @@
   "No pending approvals": "No pending approvals",
   "When users request approval for sensitive flows, they will appear here.": "When users request approval for sensitive flows, they will appear here.",
   "Approved": "Approved",
-  "Add a description": "Add a description",
   "Add knowledge": "Add knowledge",
   "Add tool": "Add tool",
   "Agent actions": "Agent actions",
@@ -50,16 +49,13 @@
   "Describe a task for your agent… e.g. research our competitors and send me a weekly brief": "Describe a task for your agent… e.g. research our competitors and send me a weekly brief",
   "Different account per action": "Different account per action",
   "Account was deleted": "Account was deleted",
-  "Edit name and appearance": "Edit name and appearance",
   "Enrich a new lead with company details and write the first email": "Enrich a new lead with company details and write the first email",
   "It may have been deleted by someone else on the project.": "It may have been deleted by someone else on the project.",
   "It may have been deleted, or the connection dropped.": "It may have been deleted, or the connection dropped.",
   "Its instructions, its tools, and every conversation held with it are deleted for good. Any draft flow step using it will break.": "Its instructions, its tools, and every conversation held with it are deleted for good. Any draft flow step using it will break.",
   "Its instructions, its tools, and every conversation held with it go with it.": "Its instructions, its tools, and every conversation held with it go with it.",
   "Knowledge": "Knowledge",
-  "knowledgeSourcesCount": "{count, plural, =1 {1 source} other {# sources}}",
   "Lead enrichment": "Lead enrichment",
-  "Needs a model": "Needs a model",
   "New agent": "New agent",
   "New agents go to": "New agents go to",
   "Pick where this agent should live.": "Pick where this agent should live.",
@@ -94,16 +90,9 @@
   "That agent could not be deleted.": "That agent could not be deleted.",
   "That agent could not be loaded": "That agent could not be loaded",
   "That agent is gone": "That agent is gone",
-  "toolsAddedCount": "{count, plural, =1 {1 tool} other {# tools}}",
   "Turn on a provider for chat": "Turn on a provider for chat",
-  "Write instructions before testing": "Write instructions before testing",
-  "Try it before it goes live": "Try it before it goes live",
-  "Needs a model to run": "Needs a model to run",
   "Try “Only reply to paying customers” or “Add Slack and Notion”.": "Try “Only reply to paying customers” or “Add Slack and Notion”.",
-  "Give it a real task. It runs on the settings beside you.": "Give it a real task. It runs on the settings beside you.",
   "Test": "Test",
-  "Pick a model before testing": "Pick a model before testing",
-  "Try {name}...": "Try {name}...",
   "Give the agent a real task. It answers on the settings beside you, so you can watch a change work before every flow gets it.": "Give the agent a real task. It answers on the settings beside you, so you can watch a change work before every flow gets it.",
   "Write one by hand instead": "Write one by hand instead",
   "Your changes couldn't be staged for testing. Try again.": "Your changes couldn't be staged for testing. Try again.",
@@ -2804,12 +2793,8 @@
   "Nothing changed on your platform. Try again \u2014 if it fails a second time, contact support and we'll turn it on manually.": "Nothing changed on your platform. Try again \u2014 if it fails a second time, contact support and we'll turn it on manually.",
   "Try again": "Try again",
   "Contact support": "Contact support",
-  "Edit with AI": "Edit with AI",
   "Activepieces AI": "Activepieces AI",
   "Close Edit with AI": "Close Edit with AI",
-  "Message the builder...": "Message the builder...",
-  "Back to the agent": "Back to the agent",
-  "Describe a change": "Describe a change",
   "Adjust the instructions, add or remove tools, or change the model. For example: “Only reply to paying customers” or “Add Slack and Notion”.": "Adjust the instructions, add or remove tools, or change the model. For example: “Only reply to paying customers” or “Add Slack and Notion”.",
   "Needs a model before it can run": "Needs a model before it can run",
   "Live — flows using this agent run these settings": "Live — flows using this agent run these settings",
@@ -2818,7 +2803,6 @@
   "These edits have not gone live yet. Leave now and they are discarded.": "These edits have not gone live yet. Leave now and they are discarded.",
   "Keep editing": "Keep editing",
   "Not live yet": "Not live yet",
-  "Changes not live yet": "Changes not live yet",
   "These changes are not live. Publish to hand them to this agent and every flow using it.": "These changes are not live. Publish to hand them to this agent and every flow using it.",
   "Service account JSON": "Service account JSON",
   "Google Cloud project ID": "Google Cloud project ID",
