+++ packages/web/test/app/routes/agents/id/leave-dialog.test.tsx
@@ -2,7 +2,7 @@
 import { cleanup, render, screen } from '@testing-library/react';
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
-import { LeaveWithoutSavingDialog } from '@/app/routes/agents/id';
+import { LeaveWithoutSavingDialog } from '@/app/routes/agents/id/configure-panel';
 
 vi.mock('i18next', () => ({ t: (key: string) => key }));
