+++ packages/web/test/app/routes/agents/id/leave-guard-arming.test.tsx
@@ -81,7 +81,7 @@ vi.mock('@/features/agents/hooks/agents-hooks', () => ({
   agentsQueries: { useAgent: () => ({ data: undefined, isLoading: false }) },
 }));
 
-import { AgentConfigurePanel } from '@/app/routes/agents/id';
+import { AgentConfigurePanel } from '@/app/routes/agents/id/configure-panel';
 
 const agentWithoutModel = {
   id: 'agent_1',
